Restructure sample activities to allow for multiple FusedLocationApi examples
