+++ lost-sample/src/main/java/com/example/lost/SamplesList.java
@@ -13,4 +13,9 @@ public class SamplesList {
       new Sample(R.string.sample_settings_api_title, R.string.sample_settings_api_description,
           SettingsApiActivity.class)
   };
+
+  public static final Sample[] FUSED_API_SAMPLES = {
+      new Sample(R.string.sample_location_listener_title,
+          R.string.sample_location_listener_description, LocationListenerActivity.class)
+  };
 }
