+++ lost-sample/src/main/res/layout/activity_list_view.xml
